Termina el preprocesamiento de datos

+++ G56/machine-learning/s5/c2/desafio_evaluado/helpers.py
@@ -62,7 +62,7 @@ def grid_plot_batch(df, cols, plot_type):
 
     """
     # calcular un aproximado a la cantidad de filas
-    rows = np.ceil(df.shape[1] / cols)
+    rows = int(np.ceil(df.shape[1] / cols))
 
     # para cada columna
     for index, (colname, serie) in enumerate(df.iteritems()):
